refactor: extract triage watch helper
